модели для fastapi
