+++ fastapi_srv/src/models/person.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+from .config_mixin import ConfigMixin
+
+
+class Person(BaseModel, ConfigMixin):
+    id: str
+    name: str
